+++ src/three-d-secure/shared/errors.js
@@ -22,6 +22,11 @@ module.exports = {
     code: 'THREEDS_NOT_ENABLED',
     message: '3D Secure is not enabled for this merchant.'
   },
+  THREEDS_CAN_NOT_USE_TOKENIZATION_KEY: {
+    type: BraintreeError.types.MERCHANT,
+    code: 'THREEDS_CAN_NOT_USE_TOKENIZATION_KEY',
+    message: '3D Secure can not use a tokenization key for authorization.'
+  },
   THREEDS_HTTPS_REQUIRED: {
     type: BraintreeError.types.MERCHANT,
     code: 'THREEDS_HTTPS_REQUIRED',